Move goToSleep and button-only sleep into sleep_session
Relocate deep-sleep entry helpers out of bl.cpp. Preserve X
touchbar_prepare_for_sleep and iqs323_task_deinit ordering from #508.
Public goToSleep remains available via bl.h for touchbar_actions.
Export submitStoredLogs for the sleep path. Behavior unchanged.

Verified: pio run -e trmnl
